RocketChat/Docker/python-docker/requirements.txt
fastapi==0.111.0
paho-mqtt
pymongo
influxdb
influxdb-client